+++ AddOns/Cell/CHANGELOG.md
@@ -1,5 +1,5 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r213-release...3a434975181729091a9c3a579a9269d19f9e2890)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r216-release...e602f935173f6534f03bbb951f113989e218da1b)
 
-- Fix Spell Request
-- Update the priority of dispel types
+- Fix duration text visibility
 - Update zhTW
+- Update snippets